requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sentence-transformers
transformers
torch
pdfplumber
PyPDF2
pandas
numpy
python-dateutil 
aiofiles
python-dotenv
